refactor: Rename domain folder from main to book
